Implement Clone for most of the types
`Clone` is not implemented for `Context` as it is more than a thin wrapper
around `cl_context`, which would need manual reference counting in order to
free resources properly when dropping the last instance. As it makes sense
to put more thought into the current idea of `Context`, I decided to not
implement `Clone` for now.

The `clone()` method of `Kernel` was renamed to `data_clone()` so that it
can be distinguished from the `Clone` implementation, that does only add
another shared reference. It was done in order to keep the `Clone` API
symmetric across all OpenCL types.

Types that don't need reference counting, `Device` and `Platform` also
implement `Copy`.