fix static linking by disabling default features in cl3 (#72)
By default, the `opencl3` crate activates the `dynamic` feature flag of the `cl3` crate, which overrides the `static` flag. This means that it's impossible to do static linking with `opencl3`.

This patch adds `default-features = true` to the `cl3` dep entry, which disables the `dynamic` flag unless the `opencl3` crate's `dynamic` flag is turned on.